+++ apps/service/src/pages/home/components/hero-section.tsx
